Improved Facilitate Extension by considering invocations should be in added methods and etc
log-NewRM041 :

For added statements invocations are further checks.
invocation to added method
invocation with expression of a declared variable
etc
And  invocations  more than extracted method  should not be subsumed by extracted method to consider an statement facilitate extension.

protected and private extracted methods are not considered for backward compatibility.